+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Handle `{:safe, iolist()}` tuples in Processor and Template (for content coming from Phoenix.View).
   Expose `Template.html_concat` as potentially useful helper.
 - Reset navigation history after each print job to avoid leaking information.
+- Create new empty browser context for each target (similar to incognito tab).
 
 ## [0.3.0] - 2020-03-30